+++ src/types/mod.rs
@@ -23,9 +23,11 @@ use concordium_base::{
         Versioned,
     },
     contracts_common::{Duration, EntrypointName, Parameter},
-    encrypted_transfers,
-    encrypted_transfers::types::{
-        AggregatedDecryptedAmount, EncryptedAmountTransferData, SecToPubAmountTransferData,
+    encrypted_transfers::{
+        self,
+        types::{
+            AggregatedDecryptedAmount, EncryptedAmountTransferData, SecToPubAmountTransferData,
+        },
     },
     id::{
         constants::{ArCurve, AttributeKind},
@@ -34,6 +36,7 @@ use concordium_base::{
             AccountAddress, AccountCredentialWithoutProofs, AccountKeys, CredentialPublicKeys,
         },
     },
+    protocol_level_tokens::{TokenEvent, TokenEventDetails, TokenHolder},
     smart_contracts::{
         ContractEvent, ModuleReference, OwnedParameter, OwnedReceiveName, WasmVersion,
     },
@@ -928,6 +931,12 @@ pub enum SpecialTransactionOutcome {
 }
 
 impl SpecialTransactionOutcome {
+    /// Return the list of addresses affected by the SpecialTransactionOutcome.
+    /// These are addresses that have their CCD balance
+    /// changed by rewards (validation_rewards, mint_rewards, block_rewards,
+    /// finalization_rewards and payday rewards). As well as validator addresses
+    /// that have been suspended/primed for suspension as part of the
+    /// SpecialTransactionOutcome.
     pub fn affected_addresses(&self) -> Vec<AccountAddress> {
         match self {
             SpecialTransactionOutcome::BakingRewards { baker_rewards, .. } => {
@@ -1120,6 +1129,7 @@ impl BlockItemSummary {
             BlockItemSummaryDetails::AccountTransaction(ad) => ad.is_rejected().is_none(),
             BlockItemSummaryDetails::AccountCreation(_) => true,
             BlockItemSummaryDetails::Update(_) => true,
+            BlockItemSummaryDetails::TokenCreationDetails(_) => true,
         }
     }
 
@@ -1135,6 +1145,7 @@ impl BlockItemSummary {
             BlockItemSummaryDetails::AccountTransaction(ad) => ad.is_rejected(),
             BlockItemSummaryDetails::AccountCreation(_) => None,
             BlockItemSummaryDetails::Update(_) => None,
+            BlockItemSummaryDetails::TokenCreationDetails(_) => None,
         }
     }
 
@@ -1143,6 +1154,7 @@ impl BlockItemSummary {
             BlockItemSummaryDetails::AccountTransaction(at) => Some(at.sender),
             BlockItemSummaryDetails::AccountCreation(_) => None,
             BlockItemSummaryDetails::Update(_) => None,
+            BlockItemSummaryDetails::TokenCreationDetails(_) => None,
         }
     }
 
@@ -1237,6 +1249,8 @@ impl BlockItemSummary {
     }
 
     /// Return the list of addresses affected by the block summary.
+    /// These are addresses that have their CCD balance or plt token balance
+    /// changed as part of the block summary.
     pub fn affected_addresses(&self) -> Vec<AccountAddress> {
         if let BlockItemSummaryDetails::AccountTransaction(at) = &self.details {
             match &at.effects {
@@ -1304,8 +1318,35 @@ impl BlockItemSummary {
                 AccountTransactionEffects::DataRegistered { .. } => vec![at.sender],
                 AccountTransactionEffects::BakerConfigured { .. } => vec![at.sender],
                 AccountTransactionEffects::DelegationConfigured { .. } => vec![at.sender],
-                AccountTransactionEffects::TokenHolder { .. } => vec![at.sender],
-                AccountTransactionEffects::TokenGovernance { .. } => vec![at.sender],
+                AccountTransactionEffects::TokenHolder { events }
+                | AccountTransactionEffects::TokenGovernance { events } => {
+                    let mut addresses = BTreeSet::new();
+                    addresses.insert(at.sender);
+
+                    for token_event in events {
+                        match &token_event.event {
+                            TokenEventDetails::Module(_) => {
+                                // An `address` added/removed from an allow/deny
+                                // list is NOT considered affected.
+                            }
+
+                            TokenEventDetails::Transfer(event) => {
+                                let TokenHolder::HolderAccount(from) = &event.from;
+                                addresses.insert(from.address);
+
+                                let TokenHolder::HolderAccount(to) = &event.to;
+                                addresses.insert(to.address);
+                            }
+
+                            TokenEventDetails::Mint(event) | TokenEventDetails::Burn(event) => {
+                                let TokenHolder::HolderAccount(to) = &event.target;
+                                addresses.insert(to.address);
+                            }
+                        }
+                    }
+
+                    addresses.into_iter().collect()
+                }
             }
         } else {
             Vec::new()
@@ -1771,6 +1812,8 @@ pub enum BlockItemSummaryDetails {
     /// The summary is of a chain update, and the outcome is as specified by the
     /// payload.
     Update(UpdateDetails),
+    /// The summary is of a token creation update.
+    TokenCreationDetails(TokenCreationDetails),
 }
 
 #[derive(Debug, Clone)]
@@ -2201,6 +2244,15 @@ pub struct UpdateDetails {
 
 impl UpdateDetails {
     pub fn update_type(&self) -> UpdateType { self.payload.update_type() }
+}
+
+#[derive(Debug, Clone)]
+// Details about the creation of a protocol-level token.
+pub struct TokenCreationDetails {
+    // The update payload used to create the token.
+    pub create_plt: CreatePlt,
+    // The events generated by the token module during the creation of the token.
+    pub events:     Vec<TokenEvent>,
 }
 
 #[derive(SerdeSerialize, SerdeDeserialize, Debug, Clone)]