+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+- Extend `BlockItemSummaryDetails` with `TokenCreationDetails` variant and handle `create plt` event.
+- Change JSON serialization of PLT events to align them with Haskell code base.
+- Expanded the `affected_addresses` function within the `BlockItemSummary` implementation to return a vector of addresses whose CCD or PLT token balances were impacted by the transaction.
+
 ## 7.0.0-alpha
 
 - Protocol level token events and reject reasons are now defined in `concordium_base::protocol_level_tokens`.